moved a lot of files around. Added a roadmap for V 1.0 release started working on the master test bench. I got the rough idea out, but its gonna take some work to get the TEST BENCH debugged I might need a simulation wishbone slave, might be able to simulate the wishbone simulation slave right in the test bench... that might be the best approach becuase I can see the signals going in and out
need to figure out how to write a Makefile that will conditionally call all the simulations if I wanted to, as well as conditionally compile.
It is the goal that I can just run "make sim" and watch the test results come back, if there is a failure then I can go there and fix it.