vetted a design that included the "uart_io_handler.v", "wishbone_master.v", "simple_gpio.v", and "drt.v" on an real FPGA, the designed worked! :)
changing the way in which the master_handlers will be written. There should be one monolithic file that controls all IO functions, this will make it easier for
the gen_top script to connect things together while the actual handler wraps up the phy

cleaned up a bunch of HDL that icaraus will compile, but Xilinx would not.

starting to work on the "gen_top.v" script that will programatically generate the top module.