Should have commited a long time ago. the code is generating a project directory but only through a test bench
The steps to generate a project is to
1. go to <path>/sycamore/sap/saplib
2. 'make gen_modules'
3. edit '<path>/sycamore/sap/saplib/data/example_projects/example1.json' to whatever you want
4. 'make test_project'

Currently working on both a tft driver and the python code to exercise it. The code already can excercise the "simple_gpio.v" module, but I have to fix the ucf file so the it doesn't need to be edited everytime you want to use a different set of headers.

Everything seems to work fine on the spartan3e starter kit board from digikey