fixed the arbitrator generated arbitrator bugs
unfortunately the arbitrator I wrote by hand (wishbone_arbitrator.v) had
an error in it that returned false postives so even though I though that
the simulation was correct all the arbitration was wrong.

The main problem I noticed was that when the master receives an ack it puts down
the stb pin, which is what the arbitrators used to detect if a specific master
was attempting to access the slave, so when it went down, the slave didn't have
a chance to respond to a strobe down. Unfortunately the original code set all
signals to x's which didn't lower the stb pin either. So what I is put all the
signals to zero when a master is deselected, now the slave should always go
back to a default reset state (stb -> 0)

now I gotta try it again on the FPGA