modified the uart protocol to send back the count of data
-Although this makes a lot more overhead for sending data
I felt it was necessary due to the nature of the drivers.
The Linux side drivers are disjointed when sending and
receiving data, so the receiver needs to know how much data is
to be read back from Sycamore. One method is to constantly by
communicating with the transmitter, but that could lead to a lot
of problems, especially down the road when the code gets ported
over to PCIe and we try and support 'relaxed ordering'

The UART protocol was never meant for high speed, most for debugging
and low speed protocols.

I have noticed some communicationg bugs with regards to the UART.
It might be due to a bug that was just fixed within the
uart_io_handler.v, so far afte the fix, everything seems to be working
well, but The bug may still linger. It could have to do with not
explicitely declaring an clock buffer.

On the driver side, I have managed to grab the data comming in from
the UART serial port directly from the URB, so the user can still
interact with sycamore through the serial port but the sycamore
driver will also get the data be able to send it to the
virtual devices.