started working on a uart protocol layer
I'm still refining the way in which the protocols interact
with each other, and I believe that the protocol layer should
only handle

ping, read, write and interrupts.

The sycamore bus layer should be the one that sends and receives
DRT request.

ping and interrupts will be easy but read and write will be tricky
only because it could require multiple transactions.