Added the function to get a list of available slaves
this originally was in the saplib but I changed this to be in the
saputily. the function searches through the slaves and finds all
verilog files, after that it scans for the meta data tags to
determine if the verilog file is a wishbone slave

I was working with a slave palette that would appear on the right
hand side of the wishbone bus view. I used an icon view that allowed
me to just use a simple icon to display the data but the icons were
not modifiable so it didn't look correctly. I might be able to
implement my own icons but it might just be easier to implement another
drawing canvas in a scrolled window