modified the way board and configuration data is read in
in general when a board is defined in the sycamore/miracle_grow/board
directory the following components are also tied to that board

-FPGA part number
-reset is inverted or not
-constraint files
-build tool (such as xilinx or altera)

the way the configuration files were set up before the user would
be required to know all of these configuration settings

that was a easy place to make mistakes.

By putting the configuration data within the board/<board name> config.json
file the developer of the board can define these configuration values once
and then the users will never have to worry about that.

In the future I would like to add support for open source toolchains
but that feels a long way off, but when that comes I will have to change
the config file to use the open source toolchain

The following components of any board/<name>/config.json file are
	-board_name: (string) defines the readible board name like "Sycamore 1"
	-vendor: (string) defines the vendor of the board like "Cospan Design"
	-fpga_part_number: (string) defines the pn of the FPGA
	-build_tool: (string) defines the toolchain to use like "xilinx"
	-constraint_files: (string list) defines all the constraint files
		associated with this board
	-invert_reset: (boolean) defines if the reset pin is inverted